backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.3.2
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
twilio==9.0.0
google-generativeai==0.8.0
anthropic>=0.45.0
deepgram-sdk==3.5.0
elevenlabs==1.5.0
deep-translator==1.11.4
langdetect==1.0.9
apscheduler==3.10.4
Pillow==10.4.0
python-multipart==0.0.9
pyngrok==7.1.6
pytest==8.3.5
